Sprint Challenge
I have gotten everythign to be responsive to an extant, I ended up making a more fluid page than adaptive. I still need to work on getting the lambda logo to go to the top at 700px, and also getting the boxes over the bnav bar at the top when the page is at 400px.