Sprint Challenge
I just can not get the lambda logo to go to the top, and the boxes to work on the nav bar at 400px. I ended up changing the width of 800px to 100% in the beginning because of problmes I was having and could not figure out. Which this move ended up being a mistake for the end goal, although the site goes to the forms of the certain screen it changes with every decrease or increase in page size.